d0_simple_langchain/requirements.txt
Flask==3.0.0
openai==1.6.1
python-dotenv==1.0.0
langchain-openai==0.0.2
langchain-core==0.1.4